+++ braindecode/augmentation/base.py
@@ -6,7 +6,7 @@
 # License: BSD (3-clause)
 
 from numbers import Real
-from typing import Any, Callable, List, Optional, Tuple, Union
+from typing import Any, Callable, Optional, Union
 
 import torch
 from sklearn.utils import check_random_state
@@ -16,17 +16,17 @@ from torch.utils.data._utils.collate import default_collate
 
 from .functional import identity
 
-Batch = List[Tuple[torch.Tensor, int, Any]]
+Batch = list[tuple[torch.Tensor, int, Any]]
 Output = Union[
     # just outputting X
     torch.Tensor,
     # outputting (X, y) where y can be a tensor or tuple of tensors
-    Tuple[torch.Tensor, Union[torch.Tensor, Tuple[torch.Tensor, ...]]],
+    tuple[torch.Tensor, Union[torch.Tensor, tuple[torch.Tensor, ...]]],
 ]
 # (X, y) -> (X', y') where y' can be a tensor or a tuple of tensors
 Operation = Callable[
     [torch.Tensor, torch.Tensor],
-    Tuple[torch.Tensor, Union[torch.Tensor, Tuple[torch.Tensor, ...]]],
+    tuple[torch.Tensor, Union[torch.Tensor, tuple[torch.Tensor, ...]]],
 ]