+++ braindecode/models/eegnet.py
@@ -3,19 +3,20 @@
 # License: BSD (3-clause)
 from __future__ import annotations
 
-from typing import Dict, List, Optional
+from typing import Dict, Optional
 
 from einops.layers.torch import Rearrange
 from mne.utils import warn
 from torch import nn
 
-from braindecode.functional import glorot_weight_zero_bias, squeeze_final_output
+from braindecode.functional import glorot_weight_zero_bias
 from braindecode.models.base import EEGModuleMixin
 from braindecode.modules import (
     Conv2dWithConstraint,
     Ensure4d,
     Expression,
     LinearWithConstraint,
+    SqueezeFinalOutput,
 )
 
 
@@ -100,9 +101,9 @@ class EEGNetv4(EEGModuleMixin, nn.Sequential):
         final_layer_with_constraint: bool = False,
         norm_rate: float = 0.25,
         # Other ways to construct the signal related parameters
-        chs_info: Optional[List[Dict]] = None,
-        input_window_seconds: Optional[float] = None,
-        sfreq: Optional[float] = None,
+        chs_info: Optional[list[Dict]] = None,
+        input_window_seconds=None,
+        sfreq=None,
         **kwargs,
     ):
         super().__init__(
@@ -283,7 +284,7 @@ class EEGNetv4(EEGModuleMixin, nn.Sequential):
                 Rearrange("batch x y z -> batch x z y"),
             )
 
-            module.add_module("squeeze", Expression(squeeze_final_output))
+            module.add_module("squeeze", SqueezeFinalOutput())
         else:
             module.add_module("flatten", nn.Flatten())
             module.add_module(
@@ -465,7 +466,7 @@ class EEGNetv1(EEGModuleMixin, nn.Sequential):
             Rearrange("batch x y z -> batch x z y"),
         )
 
-        module.add_module("squeeze", Expression(squeeze_final_output))
+        module.add_module("squeeze", SqueezeFinalOutput())
 
         self.add_module("final_layer", module)