+++ braindecode/modules/filter.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from functools import partial
-from typing import List, Optional, Tuple
+from typing import Optional
 
 import torch
 from einops.layers.torch import Rearrange
@@ -35,7 +35,7 @@ class FilterBankLayer(nn.Module):
         Number of channels in the input signal.
     sfreq : int
         Sampling frequency of the input signal in Hz.
-    band_filters : Optional[List[Tuple[float, float]]] or int, default=None
+    band_filters : Optional[list[tuple[float, float]]] or int, default=None
         List of frequency bands as (low_freq, high_freq) tuples. Each tuple defines
         the frequency range for one filter in the bank. If not provided, defaults
         to 9 non-overlapping bands with 4 Hz bandwidths spanning from 4 to 40 Hz.
@@ -128,8 +128,8 @@ class FilterBankLayer(nn.Module):
     def __init__(
         self,
         n_chans: int,
-        sfreq: int,
-        band_filters: Optional[List[Tuple[float, float]] | int] = None,
+        sfreq: float,
+        band_filters: Optional[list[tuple[float, float]] | int] = None,
         method: str = "fir",
         filter_length: str | float | int = "auto",
         l_trans_bandwidth: str | float | int = "auto",
@@ -239,11 +239,6 @@ class FilterBankLayer(nn.Module):
 
         self.filts = nn.ParameterDict(filts)
 
-        if self.method_iir:
-            self._apply_filter_func = self._apply_iir
-        else:
-            self._apply_filter_func = partial(self._apply_fir, n_chans=self.n_chans)
-
     def forward(self, x: torch.Tensor) -> torch.Tensor:
         """
         Apply the filter bank to the input signal.
@@ -258,10 +253,21 @@ class FilterBankLayer(nn.Module):
         torch.Tensor
             Filtered output tensor of shape (batch_size, n_bands, n_chans, filtered_time_points).
         """
-        return torch.cat(
-            [self._apply_filter_func(x, p_filt) for p_filt in self.filts.values()],
-            dim=1,
-        )
+        filtered_x: list[torch.Tensor] = []
+        # as we have few filters, it is okay to use this for loop...abs
+
+        for key, p_filt in self.filts.items():
+            if self.method_iir:
+                # Apply the filter to the input tensor
+                filtered = self._apply_iir(x, p_filt)
+            else:
+                # Apply the filter to the input tensor
+                filtered = self._apply_fir(x, p_filt, self.n_chans)
+
+            # Append the filtered tensor to the list
+            filtered_x.append(filtered)
+
+        return torch.cat(filtered_x, dim=1)
 
     @staticmethod
     def _apply_fir(x, filter: dict, n_chans: int) -> Tensor: